requirements.txt
django-rest-registration==0.7.0
Django==4.0.6
djangorestframework==3.13.1
python-decouple==3.6
dj-database-url==0.5.0
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
whitenoise==6.2.0
django-cors-headers==3.13.0
